Moved things into folders to make searching for some things easier, renamed CannyDetector to visionProcessing (for claritys sake)
